Feat: move svg components into new folder called Icons and update imports #11
