+++ src/lib/components/icons/Favicon.svelte
