+++ src/lib/components/icons/Gallery.svelte
