+++ src/lib/components/icons/Grid3.svelte
