+++ src/lib/components/icons/Grid5.svelte
