+++ src/lib/components/icons/Group.svelte
