+++ src/lib/components/icons/Heart.svelte
