+++ src/lib/components/icons/Listview.svelte
