+++ src/lib/components/icons/Logo.svelte
