+++ src/lib/components/icons/Search.svelte
