+++ src/lib/components/icons/Star.svelte
